Refactor: 공유페이지 api 함수를 다른 파일로 빼냄 - #31
