+++ src/lib/api/index.ts
@@ -2,3 +2,4 @@ export * from './auth';
 export * from './config';
 export * from './folder';
 export * from './user';
+export * from './share';
