+++ src/pages/shared/[id]/index.tsx
@@ -1,37 +1,12 @@
 // TODO: 파일 내에 있는 주석들 수정 예정
 import { GetStaticPaths, GetStaticProps } from 'next';
 // import { useContext } from 'react';
-import { API_PATH, FolderTypes, GetFolderParams } from '@/lib/api';
 // import { FolderContext } from '@/lib/context';
 import CardList from '@/components/favorite/CardList/CardList';
-import axios from 'axios';
-
-const tempInstance = axios.create({
-  baseURL: process.env.NEXT_PUBLIC_BASE_URL,
-  timeout: 3000,
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
-
-const getFolderList = async () => {
-  const response = await tempInstance.get<FolderTypes[]>(
-    API_PATH.folder.default,
-  );
-
-  return response.data ?? [];
-};
-
-const getLinkList = async ({ folderId }: GetFolderParams) => {
-  const response = await tempInstance.get<FolderTypes[]>(
-    API_PATH.link.category(folderId),
-  );
-
-  return response.data ?? [];
-};
+import { getFolderListForSSG, getLinkListForSSG } from '@/lib/api';
 
 export const getStaticPaths: GetStaticPaths = async () => {
-  const folderList = await getFolderList();
+  const folderList = await getFolderListForSSG();
 
   const paths = folderList.map((folder) => ({
     params: { id: folder.id.toString() },
@@ -45,7 +20,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
 
 export const getStaticProps: GetStaticProps = async (context) => {
   const folderId = parseInt(context.params?.id as string, 10);
-  const linkList = await getLinkList({ folderId });
+  const linkList = await getLinkListForSSG({ folderId });
 
   return {
     props: {
